requirements_qt.txt
PySide6
matplotlib
numpy
scipy
pandas
openpyxl
pillow
